requirements.txt
# IA & NLP
google-generativeai>=0.3.0
nltk>=3.8.0

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Utilitários
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2==3.0.1
unidecode>=1.3.0

# Científicos básicos
numpy>=1.24.0
scikit-learn>=1.3.0
